lfs: `track`/`untrack` support sub-dirs, but all patterns record in root attributes file (to improve)
