+++ libra/src/utils/mod.rs
@@ -4,3 +4,4 @@ pub(crate) mod path;
 pub(crate) mod object_ext;
 pub(crate) mod path_ext;
 pub(crate) mod client_storage;
+pub(crate) mod lfs;
